webservices/serve_iris_predictions/requirements.txt
Flask==1.0.2
google-cloud-storage
gpt_2_simple
sklearn
joblib
